Simplied enterprise_enrolments by modularizing tables
